Temporarily remove math from CI
The current math tests basically don't do anything as their value
are hardcoded meaning the source of truth is itself. Any actual
changes to the math code will make the build fail, then we'll go
change the numbers, essentially destroying the point of the test.

Instead, in the future, we should probably have the rust
sdk (which uses the actual contract code) spit out a testing
file and python code could run against that. This is called
a parity test. For now, i'll comment out the math.